fix: clear the 11 strict-mode warnings in the public headers
JUCE's CMake applies juce_recommended_warning_flags by default, so every
integrator compiled the SDK headers under -Wextra-class settings and got 11
Clang warnings out of code they do not own. Anyone building with -Werror had
to carve include/ out of it.

All three fixes are behaviour-neutral:

  - PersistData had four fields and three call sites passed three, so
    license_key was default-initialised. nullopt was already the intended
    value — persist_ reads it as "keep the cached value" — so the omission
    was correct, just implicit. Spelled it, and generalised the struct comment
    so the contract covers every field rather than only lease_json.

  - Seven signed loop indices subscripted containers whose operator[] takes an
    unsigned size_type, plus one int subtracted from a size_t length. Indices
    are now size_t; the length subtraction is an explicit cast.

  - apply_offline_grace_ took a `State before` it never read: every branch
    re-reads state_.load() instead, which is the behaviour we want, since the
    state can move while the network call is in flight. Dropped the parameter
    and the now-unused local at its only caller rather than freeze a stale
    snapshot into the signature.

Verified: integrations/juce/citest against a real JUCE 8.0.6 checkout now
builds with zero warnings and the smoke test passes; the SDK suite and the
amalgamation test pass; keylight_single.hpp regenerated.

The JUCE README's verification step said to expect these warnings — updated
to say a clean build is the pass criterion, which it now is.

+++ include/keylight/ed25519.hpp
@@ -117,7 +117,7 @@ static void kl_sha512_update(KlSha512Ctx& ctx, const kl_u8* data, size_t len) {
         int space = 128 - ctx.used;
         int take  = (len < (size_t)space) ? (int)len : space;
         for (int i = 0; i < take; ++i) ctx.buf[ctx.used + i] = data[i];
-        ctx.used += take; data += take; len -= take;
+        ctx.used += take; data += take; len -= static_cast<size_t>(take);
         if (ctx.used == 128) {
             kl_sha512_block(ctx.h, ctx.buf);
             ctx.total += 128; ctx.used = 0;
@@ -452,8 +452,8 @@ inline bool ed25519_verify(const uint8_t* msg, size_t msg_len,
     // h = SHA-512(R ‖ A ‖ M) mod L
     // where R = sig[0..31], A = pubkey[0..31].
     kl_u8 prefix[64];
-    for (int i = 0; i < 32; ++i) prefix[i]    = sig[i];
-    for (int i = 0; i < 32; ++i) prefix[32+i] = pubkey[i];
+    for (size_t i = 0; i < 32; ++i) prefix[i]    = sig[i];
+    for (size_t i = 0; i < 32; ++i) prefix[32+i] = pubkey[i];
 
     KlSha512Ctx sha_ctx;
     kl_sha512_init(sha_ctx);
